ok, i've added search functionality.  basically you SetPerlVar "EnableSearch" to a truevalue.  this causes Apache::MP3 to iterate over "BaseDir", opening all mp3 files and caching the info from the ID3 (or WAV/Vorbis equivalent) in a big hash.
the search function appears by default in the upper-right corner, someone
can set up code to configure where it appears, if they like.

the way i'm caching the id3 tags depends on url-uri conversion to find the
base mp3 dir, and uri-url conversion to save it in a manner that is
streamable.  i'm not entirely sure if this is compatible with symbolic
links.  please test it.

-allen